move MethodConfig.createFlattening to FlatteningConfig
